services/ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = ""
authors = [
    {name = "plutopulp",email = "plutopulped@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "readabilipy (>=0.3.0,<0.4.0)",
    "httpx (>=0.27.0,<1.0.0)",
    "langchain-openai (>=0.3.35,<0.4.0)",
    "langchain-anthropic (>=0.3.22,<0.4.0)",
    "nicknames (>=1.0.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
isort = "^7.0.0"
black = "^25.9.0"
pytest = "^8.4.2"

[tool.isort]
profile = "black"
